refactor: service test repository comment 분리
